Added reset logic
